+++ packages/syft/src/syft/client/domain_client.py
@@ -8,7 +8,6 @@ from typing import TYPE_CHECKING
 from typing import cast
 
 # third party
-from hagrid.orchestra import NodeHandle
 from loguru import logger
 from tqdm import tqdm
 
@@ -43,6 +42,7 @@ from .protocol import SyftProtocol
 
 if TYPE_CHECKING:
     # relative
+    from ..orchestra import NodeHandle
     from ..service.project.project import Project
